30.03.2022

+++ ft_memset.c
@@ -11,8 +11,7 @@
 /* ************************************************************************** */
 
 
-#include <stdlib.h>
-void *ft_memset(void *b, int c, size_t len)
+void *ft_memset(void *b, int c, unsigned int len)
 {
     unsigned int i;
